+++ build.gradle
@@ -22,5 +22,6 @@ test {
         suites "src/test/resources/HomeWork19.xml"
         suites "src/test/resources/HomeWork20.xml"
         suites "src/test/resources/HomeWork21.xml"
+        suites "src/test/resources/HomeWork22.xml"
     }
 }
