+++ src/test/java/Homework19.java
@@ -1,8 +1,7 @@
-import org.apache.hc.core5.reactor.Command;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.LoginPage;
-import pages.PlayList;
+import pages.HomePage;
+import pages.PlayListPage;
 
 public class Homework19 extends BaseTest {
 
@@ -11,11 +10,12 @@ public class Homework19 extends BaseTest {
 
     @Test(priority = 1)
     public void deletePlaylist() {
-        PlayList playList = new PlayList(driver);
-        playList.deletePlayListUsingRedButton();
+        HomePage homePage = new HomePage(driver);
+        PlayListPage playListPage = new PlayListPage(driver);
+        playListPage.deletePlayListUsingRedButton();
         String expectedAlertMessage = "Deleted playlist \"MyPlayList.\"";
         //Assert.assertTrue(playList.),);
-        Assert.assertEquals(playList.successfulDeletionMessage(),expectedAlertMessage,"Assertion failed : User created playList exists under Playlist");
+        Assert.assertEquals(homePage.successfulDeletionMessage(),expectedAlertMessage,"Assertion failed : User created playList exists under Playlist");
 
     }
