+++ src/test/resources/HomeWork22.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
+
+<suite name="HomeWork22-Suite">
+    <parameter name="baseUrl" value="https://qa.koel.app/"></parameter>
+    <test name="Homework22 Test" preserve-order="false">
+        <classes>
+            <class name="Homework22"/>
+
+
+        </classes>
+    </test>
+</suite>
